almost no ext

+++ src/integrator_ias15.c
@@ -312,7 +312,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p1[k];
 						const double gk = at[k] - a0[k];
-						g.p1[k] = (gk/rr[1] - g.p0[k])/rr[2];
+						g.p1[k] = (gk/rr[1] - (double)g.p0[k])/rr[2];
 						tmp = g.p1[k] - tmp;
 						b.p0[k] += tmp * c[0];
 						b.p1[k] += tmp;
@@ -321,7 +321,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p2[k];
 						const double gk = at[k] - a0[k];
-						g.p2[k] = ((gk/rr[3] - g.p0[k])/rr[4] - g.p1[k])/rr[5];
+						g.p2[k] = ((gk/rr[3] - (double)g.p0[k])/rr[4] - (double)g.p1[k])/rr[5];
 						tmp = g.p2[k] - tmp;
 						b.p0[k] += tmp * c[1];
 						b.p1[k] += tmp * c[2];
@@ -331,7 +331,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p3[k];
 						const double gk = at[k] - a0[k];
-						g.p3[k] = (((gk/rr[6] - g.p0[k])/rr[7] - g.p1[k])/rr[8] - g.p2[k])/rr[9];
+						g.p3[k] = (((gk/rr[6] - (double)g.p0[k])/rr[7] - (double)g.p1[k])/rr[8] - (double)g.p2[k])/rr[9];
 						tmp = g.p3[k] - tmp;
 						b.p0[k] += tmp * c[3];
 						b.p1[k] += tmp * c[4];
@@ -342,7 +342,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p4[k];
 						const double gk = at[k] - a0[k];
-						g.p4[k] = ((((gk/rr[10] - g.p0[k])/rr[11] - g.p1[k])/rr[12] - g.p2[k])/rr[13] - g.p3[k])/rr[14];
+						g.p4[k] = ((((gk/rr[10] - (double)g.p0[k])/rr[11] - (double)g.p1[k])/rr[12] - (double)g.p2[k])/rr[13] - (double)g.p3[k])/rr[14];
 						tmp = g.p4[k] - tmp;
 						b.p0[k] += tmp * c[6];
 						b.p1[k] += tmp * c[7];
@@ -354,7 +354,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p5[k];
 						const double gk = at[k] - a0[k];
-						g.p5[k] = (((((gk/rr[15] - g.p0[k])/rr[16] - g.p1[k])/rr[17] - g.p2[k])/rr[18] - g.p3[k])/rr[19] - g.p4[k])/rr[20];
+						g.p5[k] = (((((gk/rr[15] - (double)g.p0[k])/rr[16] - (double)g.p1[k])/rr[17] - (double)g.p2[k])/rr[18] - (double)g.p3[k])/rr[19] - (double)g.p4[k])/rr[20];
 						tmp = g.p5[k] - tmp;
 						b.p0[k] += tmp * c[10];
 						b.p1[k] += tmp * c[11];
@@ -370,7 +370,7 @@ static int reb_integrator_ias15_step(struct reb_simulation* r) {
 					for(int k=0;k<N3;++k) {
 						double tmp = g.p6[k];
 						const double gk = at[k] - a0[k];
-						g.p6[k] = ((((((gk/rr[21] - g.p0[k])/rr[22] - g.p1[k])/rr[23] - g.p2[k])/rr[24] - g.p3[k])/rr[25] - g.p4[k])/rr[26] - g.p5[k])/rr[27];
+						g.p6[k] = ((((((gk/rr[21] - (double)g.p0[k])/rr[22] - (double)g.p1[k])/rr[23] - (double)g.p2[k])/rr[24] - (double)g.p3[k])/rr[25] - (double)g.p4[k])/rr[26] - (double)g.p5[k])/rr[27];
 						tmp = g.p6[k] - tmp;	
 						b.p0[k] += tmp * c[15];
 						b.p1[k] += tmp * c[16];
